Refactor grading system and update imports
